Initial commit of word generator project.
The code for the generator itself has been copied from the daily greeting generator project, jabberwocky.py in the generator module.
A main loop with clarg parsing, file input and output, and error handling was implemented.
Progress bars were added, logging was replaced with print statements, and consistent whitespace and verbosity was ensured.
In the future, this will be able to be installed as a module and used by the greeting generator as such.
Furthermore, test files will be downloaded directly from project gutenberg.